+++ src/main/java/com/fitbum/repositorios/plantillas/PlantMicroRepositorio.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlantMicroRepositorio extends JpaRepository<PlantillaMicrociclo,Long> {
+public interface PlantMicroRepositorio extends JpaRepository<PlantillaMicrociclo,Integer> {
 }
