+++ src/main/java/com/fitbum/repositorios/plantillas/PlantSesionRepositorio.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlantSesionRepositorio extends JpaRepository<PlantillaSesion,Long> {
+public interface PlantSesionRepositorio extends JpaRepository<PlantillaSesion,Integer> {
 }
