+++ src/main/java/com/fitbum/repositorios/programa/MesocicloRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MesocicloRepositorio  extends JpaRepository<Mesociclo,Long> {
+public interface MesocicloRepositorio  extends JpaRepository<Mesociclo,Integer> {
 }
